% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensr.R
\name{package_licenses}
\alias{package_licenses}
\title{Find licenses for specified packages}
\usage{
package_licenses(pkglist, sort = TRUE)
}
\arguments{
\item{pkglist}{Character. Package names.}

\item{sort}{Boolean. Return results sorted by license type? If \code{FALSE}
returns in the order of packages listed in \code{pkglist}}
}
\value{
A data frame with package names and package licenses
}
\description{
Find licenses for specified packages
}
