% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensr.R
\name{project_licenses}
\alias{project_licenses}
\title{Find licenses for all packages in a project directory}
\usage{
project_licenses(dir = ".")
}
\arguments{
\item{dir}{Project directory. Defaults to current working directory.}
}
\value{
A data frame with package names and package licenses
}
\description{
Find licenses for all packages in a project directory
}
