% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensr.R
\name{package_licenses}
\alias{package_licenses}
\title{Find licenses for specified packages}
\usage{
package_licenses(pkglist)
}
\arguments{
\item{pkglist}{Character. Package names.}
}
\value{
A data frame with package names and package licenses
}
\description{
Find licenses for specified packages
}
